refactor: redirect 헬퍼 메서드 생성
