+++ tomcat/src/main/java/org/apache/coyote/http11/HttpRequestReader.java
@@ -6,7 +6,6 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.util.HashMap;
 import java.util.Map;
-import org.apache.coyote.http11.handler.QueryParser;
 
 public class HttpRequestReader {
