+++ tomcat/src/main/java/org/apache/coyote/http11/QueryParser.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.handler;
+package org.apache.coyote.http11;
 
 import java.util.HashMap;
 import java.util.Map;
