+++ tomcat/src/main/java/org/apache/coyote/http11/handler/LoginHandler.java
@@ -1,13 +1,12 @@
 package org.apache.coyote.http11.handler;
 
 import com.techcourse.service.UserService;
-import java.util.LinkedHashMap;
 import java.util.Map;
 import org.apache.coyote.http11.HttpMethod;
 import org.apache.coyote.http11.HttpRequest;
 import org.apache.coyote.http11.HttpResourceLoader;
 import org.apache.coyote.http11.HttpResponse;
-import org.apache.coyote.http11.HttpStatus;
+import org.apache.coyote.http11.QueryParser;
 
 public class LoginHandler implements HttpHandler {
 
@@ -33,13 +32,6 @@ public class LoginHandler implements HttpHandler {
 
         UserService.login(account, password);
 
-        return redirect("/index.html");
-    }
-
-    private HttpResponse redirect(final String path) {
-        Map<String, String> headers = new LinkedHashMap<>();
-        headers.put("Location", path);
-
-        return new HttpResponse(HttpStatus.FOUND, headers, new byte[0]);
+        return HttpResponse.redirect("/index.html");
     }
 }
