+++ tomcat/src/main/java/org/apache/coyote/http11/handler/RegisterHandler.java
@@ -1,13 +1,12 @@
 package org.apache.coyote.http11.handler;
 
 import com.techcourse.service.UserService;
-import java.util.LinkedHashMap;
 import java.util.Map;
 import org.apache.coyote.http11.HttpMethod;
 import org.apache.coyote.http11.HttpRequest;
 import org.apache.coyote.http11.HttpResourceLoader;
 import org.apache.coyote.http11.HttpResponse;
-import org.apache.coyote.http11.HttpStatus;
+import org.apache.coyote.http11.QueryParser;
 import org.apache.coyote.http11.RequestLine;
 
 public class RegisterHandler implements HttpHandler {
@@ -36,13 +35,6 @@ public class RegisterHandler implements HttpHandler {
 
         UserService.register(account, email, password);
 
-        return redirect("/index.html");
-    }
-
-    private HttpResponse redirect(final String path) {
-        Map<String, String> headers = new LinkedHashMap<>();
-        headers.put("Location", path);
-
-        return new HttpResponse(HttpStatus.FOUND, headers, new byte[0]);
+        return HttpResponse.redirect("/index.html");
     }
 }
